+++ NEWS.md
@@ -4,6 +4,7 @@
  * `SDA_spatialQuery()` and `processSDA_WKT()` fully use {sf}, replacing {sp} in these contexts
  * `SDA_spatialQuery()` gains argument `byFeature` to use multiple single-feature queries and combine the results with a unique feature ID specified by `idcol` argument. This allows for specific feature intersection results without secondary spatial overlay of the polygons (https://github.com/ncss-tech/soilDB/issues/222)
  * `dbConnectNASIS()` no longer requires that the NASIS credentials option be set if the `dsn` argument is specified.
+ * Rebuilt `loafercreek`, `gopheridge` and `mineralKing` from latest `fetchNASIS()` results.
  
 # soilDB 2.6.9 (2021-12-02)
  * Replaced functionality using {plyr}/{reshape2} with {base}/{data.table}